Text Generation
In this project, I generated the text using the LSTM model. The first step is to clean the data and then to create tokens using the Keras Tokenizer. After that the sequence is generated which is basically consists of tokens assigned to different words using tokenizer. After that the last word of every line is taken as target variable and rest as the features. After the model is trained using LSTM and then the model was tested using some words.